+++ apps/meteor/client/components/Sidebar/Sidebar.tsx
@@ -1,10 +1,8 @@
-import { Sidebar as FuselageSidebar } from '@rocket.chat/fuselage';
+import { SidebarV2 as FuselageSidebar } from '@rocket.chat/fuselage';
 import type { ComponentPropsWithoutRef } from 'react';
 
 export type SidebarProps = ComponentPropsWithoutRef<typeof FuselageSidebar>;
 
-const Sidebar = (props: SidebarProps) => (
-	<FuselageSidebar {...props} role='navigation' display='flex' flexDirection='column' height='full' />
-);
+const Sidebar = (props: SidebarProps) => <FuselageSidebar {...props} />;
 
 export default Sidebar;