+++ apps/meteor/client/views/admin/moderation/ModConsoleReportDetails.tsx
@@ -1,7 +1,7 @@
 import type { IUser } from '@rocket.chat/core-typings';
-import { Tabs, TabsItem, ContextualbarHeader, ContextualbarTitle } from '@rocket.chat/fuselage';
+import { Tabs, TabsItem } from '@rocket.chat/fuselage';
 import { useStableCallback } from '@rocket.chat/fuselage-hooks';
-import { ContextualbarClose, ContextualbarDialog } from '@rocket.chat/ui-client';
+import { ContextualbarClose, ContextualbarDialog, ContextualbarHeader, ContextualbarTitle } from '@rocket.chat/ui-client';
 import { useTranslation, useRouter, useRouteParameter } from '@rocket.chat/ui-contexts';
 import { useState } from 'react';