+++ apps/meteor/server/api/v1/rooms.ts
@@ -2,6 +2,7 @@ import { FederationMatrix, MeteorError, Room, Team } from '@rocket.chat/core-ser
 import {
 	type IRoom,
 	type IRoomAbacRedaction,
+	type IMessage,
 	type IUpload,
 	type RequiredField,
 	type RoomAdminFieldsType,
@@ -419,11 +420,12 @@ const roomsSaveNotificationEndpoint = API.v1.post(
 	},
 );
 
-const saveDraftBodySchema = ajv.compile<{ rid: IRoom['_id']; draft: string }>({
+const saveDraftBodySchema = ajv.compile<{ rid: IRoom['_id']; draft: string; tmid?: IMessage['_id'] }>({
 	type: 'object',
 	properties: {
 		rid: { type: 'string', minLength: 1 },
 		draft: { type: 'string' },
+		tmid: { type: 'string', minLength: 1, pattern: '^[^.$]+$' },
 	},
 	required: ['rid', 'draft'],
 	additionalProperties: false,
@@ -450,13 +452,13 @@ const roomsSaveDraftEndpoint = API.v1.post(
 		},
 	},
 	async function action() {
-		const { rid, draft } = this.bodyParams;
+		const { rid, draft, tmid } = this.bodyParams;
 
 		if (draft.length > (settings.get<number>('Message_MaxAllowedSize') ?? 0)) {
 			return API.v1.failure('error-message-size-exceeded');
 		}
 
-		const subscription = await Subscriptions.updateDraftByRoomIdAndUserId(rid, this.userId, draft || undefined);
+		const subscription = await Subscriptions.updateDraftByRoomIdAndUserId(rid, this.userId, draft || undefined, tmid);
 		if (!subscription) {
 			throw new Meteor.Error('error-invalid-subscription', 'Invalid subscription');
 		}