+++ apps/uikit-playground/src/Components/Preview/Display/Surface/ContextualBarSurface.tsx
@@ -3,11 +3,11 @@ import {
 	Box,
 	Button,
 	ButtonGroup,
-	Contextualbar,
-	ContextualbarAction,
-	ContextualbarFooter,
-	ContextualbarHeader,
-	ContextualbarTitle,
+	ContextualbarV2,
+	ContextualbarV2Action,
+	ContextualbarV2Footer,
+	ContextualbarV2Header,
+	ContextualbarV2Title,
 } from '@rocket.chat/fuselage';
 import { Scrollbars } from 'rc-scrollbars';
 import type { ReactNode } from 'react';
@@ -15,12 +15,12 @@ import type { ReactNode } from 'react';
 export type ContextualBarSurfaceProps = { children: ReactNode };
 
 const ContextualBarSurface = ({ children }: ContextualBarSurfaceProps) => (
-	<Contextualbar>
-		<ContextualbarHeader>
+	<ContextualbarV2>
+		<ContextualbarV2Header>
 			<Avatar url='data:image/gif;base64,R0lGODlhAQABAIAAAMLCwgAAACH5BAAAAAAALAAAAAABAAEAAAICRAEAOw==' />
-			<ContextualbarTitle>Contextual Bar</ContextualbarTitle>
-			<ContextualbarAction data-qa='ContextualbarActionClose' title='Close' name='cross' />
-		</ContextualbarHeader>
+			<ContextualbarV2Title>Contextual Bar</ContextualbarV2Title>
+			<ContextualbarV2Action data-qa='ContextualbarActionClose' title='Close' name='cross' />
+		</ContextualbarV2Header>
 
 		<Box height='100%' padding='12px'>
 			<Box height='100%' display='flex' flexShrink={1} flexDirection='column' flexGrow={1}>
@@ -50,13 +50,13 @@ const ContextualBarSurface = ({ children }: ContextualBarSurfaceProps) => (
 			</Box>
 		</Box>
 
-		<ContextualbarFooter>
+		<ContextualbarV2Footer>
 			<ButtonGroup stretch>
 				<Button>Cancel</Button>
 				<Button primary>Submit</Button>
 			</ButtonGroup>
-		</ContextualbarFooter>
-	</Contextualbar>
+		</ContextualbarV2Footer>
+	</ContextualbarV2>
 );
 
 export default ContextualBarSurface;