+++ packages/livechat/postcss.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+	plugins: {
+		'autoprefixer': {},
+		'postcss-selector-not': {},
+		'postcss-flexbugs-fixes': {},
+		'postcss-logical': {},
+		'postcss-dir-pseudo-class': {},
+		'cssnano': {},
+	},
+};