+++ packages/ui-voip/src/components/Widget/Widget.tsx
@@ -1,61 +1,29 @@
 import { FocusScope } from '@react-aria/focus';
-import { Palette } from '@rocket.chat/fuselage';
-import styled from '@rocket.chat/styled';
-import type { ComponentProps, ReactNode } from 'react';
-import { useLayoutEffect } from 'react';
+import type { ReactNode } from 'react';
 
-import { DragContext } from './WidgetDraggableContext';
-import { useDraggable } from '../../hooks';
-import { type LastKnownPosition } from '../../providers/useWidgetPositionTracker';
-
-// TODO: Initial position from the draggable api instead of style props
-// TODO: A11Y
-const WidgetBase = styled('div')`
-	position: fixed;
-	padding: 0;
-	right: 2em;
-	top: 11em;
-	display: flex;
-	flex-direction: column;
-	width: 248px;
-	min-height: 128px;
-	border-radius: 4px;
-	border: 1px solid ${Palette.stroke['stroke-dark'].toString()};
-	box-shadow:
-		0px 0px 1px 0px ${Palette.shadow['shadow-elevation-2x'].toString()},
-		0px 0px 12px 0px ${Palette.shadow['shadow-elevation-2y'].toString()};
-	background-color: ${Palette.surface['surface-tint'].toString()};
-	color: ${Palette.text['font-default'].toString()};
-	z-index: 100;
-	overflow: hidden;
-`;
+import WidgetBase from './WidgetBase';
+import { useDraggableWidget } from './WidgetDraggableContext';
 
 export type WidgetProps = {
 	children: ReactNode;
-	restorePosition?: LastKnownPosition | null;
-	onChangePosition?: (position: LastKnownPosition | null) => void;
-} & ComponentProps<typeof WidgetBase>;
+	autoFocus?: boolean;
+};
 
-const Widget = ({ children, onChangePosition, restorePosition, ...props }: WidgetProps) => {
-	const [draggableRef, boundingRef, handleRef] = useDraggable({ onChangePosition, restorePosition });
-
-	useLayoutEffect(() => {
-		boundingRef(document.body);
-	}, [boundingRef]);
+const Widget = ({ children, autoFocus = true, ...props }: WidgetProps) => {
+	const draggableContext = useDraggableWidget();
 
 	return (
-		<DragContext.Provider value={{ draggableRef, boundingRef, handleRef }}>
-			<FocusScope autoFocus>
-				<WidgetBase
-					{...props}
-					ref={draggableRef}
-					role='dialog'
-					aria-labelledby='rcx-media-call-widget-title-prefix rcx-media-call-widget-title rcx-media-call-widget-caller-info'
-				>
-					{children}
-				</WidgetBase>
-			</FocusScope>
-		</DragContext.Provider>
+		<FocusScope autoFocus={autoFocus}>
+			<WidgetBase
+				{...props}
+				ref={draggableContext?.draggableRef}
+				inline={!draggableContext}
+				role='dialog'
+				aria-labelledby='rcx-media-call-widget-title-prefix rcx-media-call-widget-title rcx-media-call-widget-caller-info'
+			>
+				{children}
+			</WidgetBase>
+		</FocusScope>
 	);
 };