+++ apps/meteor/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @rocket.chat/meteor
 
+## 8.7.0-rc.4
+
+### Patch Changes
+
+- Bump @rocket.chat/meteor version.
+
+- ([#41588](https://github.com/RocketChat/Rocket.Chat/pull/41588)) Fixes audio attachments not being seekable using the progress slider
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@8.7.0-rc.4
+  - @rocket.chat/rest-typings@8.7.0-rc.4
+  </details>
+
+## 8.7.0-rc.3
+
+### Patch Changes
+
+- Bump @rocket.chat/meteor version.
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@8.7.0-rc.3
+  - @rocket.chat/rest-typings@8.7.0-rc.3
+  </details>
+
 ## 8.7.0-rc.2
 
 ### Patch Changes